requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0
requests>=2.26.0
python-dotenv>=0.19.0
tqdm>=4.62.0
python-dateutil>=2.8.2
pyyaml>=6.0
scipy>=1.7.0
bcrypt>=4.0.0
pytest>=7.0.0
boto3>=1.26.0